+++ src/about/page.tsx
@@ -3,9 +3,9 @@ import Header from "../components/layouts/header/header";
 import Hero from "../components/layouts/hero/hero";
 import Container from "../components/layouts/container/container";
 import StackAbout from "../components/elements/stackAbout/stackAbout";
-import ResponsiveImage from "../components/responsiveImage/responsiveImage";
+import ResponsiveImage from "../components/elements/responsiveImage/responsiveImage";
 import Contact from "../components/layouts/contact/contact";
-import Advertisement from "../components/advertisement/advertisement";
+import Advertisement from "../components/layouts/advertisement/advertisement";
 import Footer from "../components/layouts/footer/footer";
 import useScrollAnimation from "../components/elements/animation/useScrollAnimation";
 import useScrollAnimationStyles from "../components/elements/animation/useScrollAnimation.module.css";
