+++ src/components/elements/stackInfo/stackInfo.tsx
