+++ src/components/layouts/header/header.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import HeaderLogo from "../../headerLogo/headerLogo";
-import Nav from "../../nav/nav";
+import HeaderLogo from "../../elements/headerLogo/headerLogo";
+import Nav from "../../elements/nav/nav";
 import styles from "./header.module.css";
 import globalStyles from "@/src/globalStyle.module.css";
