+++ src/home/page.tsx
@@ -2,11 +2,11 @@ import React from "react";
 import Header from "../components/layouts/header/header";
 import Hero from "../components/layouts/hero/hero";
 import Container from "../components/layouts/container/container";
+import HomeSection from "../components/elements/homeSection/homeSection";
 import Stack from "../components/elements/stack/stack";
 import Button from "../components/elements/button/button";
-import HomeSection from "../components/homeSection/homeSection";
 import Contact from "../components/layouts/contact/contact";
-import Advertisement from "../components/advertisement/advertisement";
+import Advertisement from "../components/layouts/advertisement/advertisement";
 import Footer from "../components/layouts/footer/footer";
 import globalStyles from "@/src/globalStyle.module.css";
 import styles from "./page.module.css";
