+++ src/infomation/page.tsx
@@ -3,9 +3,9 @@ import { Link } from "react-router-dom";
 import Header from "../components/layouts/header/header";
 import Hero from "../components/layouts/hero/hero";
 import Container from "../components/layouts/container/container";
-import StackInfo from "../components/stackInfo/stackInfo";
+import StackInfo from "../components/elements/stackInfo/stackInfo";
 import Contact from "../components/layouts/contact/contact";
-import Advertisement from "../components/advertisement/advertisement";
+import Advertisement from "../components/layouts/advertisement/advertisement";
 import Footer from "../components/layouts/footer/footer";
 import globalStyles from "@/src/globalStyle.module.css";
 import styles from "./page.module.css";
